Rename abi.ts to routerAbi.ts
Aligns the filename with the exported symbol (`routerAbi`) and with
the sibling ABI files (`agentAbi.ts`, `vetro/gatewayAbi.ts`), which
already follow the `<name>Abi.ts` convention. The plain `abi.ts` was
the odd one out — easy to grep wrong and ambiguous in imports when
multiple ABIs live in the package.

Also fixes the import order in the eight call sites so `constants`
stays alphabetically before `routerAbi`.
