+++ packages/hemi-earn-actions/src/actions/wallet/requestDeposit.ts
@@ -4,8 +4,8 @@ import type { Address, TransactionReceipt, WalletClient } from 'viem'
 import { waitForTransactionReceipt, writeContract } from 'viem/actions'
 import { allowance, approve, balanceOf } from 'viem-erc20/actions'
 
-import { routerAbi } from '../../abi'
 import { getHemiEarnRouterAddress } from '../../constants'
+import { routerAbi } from '../../routerAbi'
 import type { RequestDepositEvents } from '../../types'
 import { quoteDeposit } from '../public/quoteDeposit'
